+++ examples/partial-prerendering/CHANGELOG.md
@@ -1,5 +1,18 @@
 # partial-prerendering
 
+## 2.6.8
+
+### Patch Changes
+
+- @stackframe/stack@2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.6.7
+
 ## 2.6.6
 
 ### Patch Changes