+++ packages/stack-emails/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @stackframe/stack-emails
 
+## 2.6.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.8
+  - @stackframe/stack-ui@2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.6.7
+- @stackframe/stack-ui@2.6.7
+
 ## 2.6.6
 
 ### Patch Changes