+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @stackframe/stack
 
+## 2.6.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.6.8
+  - @stackframe/stack-ui@2.6.8
+  - @stackframe/stack-sc@2.6.8
+
+## 2.6.7
+
+### Patch Changes
+
+- Bugfixes
+- Updated dependencies
+  - @stackframe/stack-sc@2.6.7
+  - @stackframe/stack-shared@2.6.7
+  - @stackframe/stack-ui@2.6.7
+
 ## 2.6.6
 
 ### Patch Changes